+++ ChangeLog.rst
@@ -26,7 +26,10 @@ Added
 - New ``DataEAM`` class with eam/fs parsing.
 - New ``potential::Base`` class (virtual interface class).
 - New ``EAM`` class.
-- Supercell has a SoA getter method.
+- ``Supercell`` has a ``SoA`` getter method.
+- New ``xize`` utility function.
+- New Hessian class.
+
 
 Changed
 ~~~~~~~